twister: move execution code out of testplan
Move all code related to execution into runner class.
